Move PicklableWithSlots to polyutils
PicklableWithSlots is now only used in sympy.polys.
