+++ sympy/core/basic.py
@@ -8,72 +8,6 @@ from compatibility import callable, reduce, cmp, iterable
 from sympy.core.decorators import deprecated
 from sympy.core.singleton import S
 
-class PicklableWithSlots(object):
-    """
-    Mixin class that allows to pickle objects with ``__slots__``.
-
-    Examples
-    --------
-
-    First define a class that mixes :class:`PicklableWithSlots` in::
-
-        >>> from sympy.core.basic import PicklableWithSlots
-
-        >>> class Some(PicklableWithSlots):
-        ...     __slots__ = ['foo', 'bar']
-        ...
-        ...     def __init__(self, foo, bar):
-        ...         self.foo = foo
-        ...         self.bar = bar
-
-    To make :mod:`pickle` happy in doctest we have to use this hack::
-
-        >>> import __builtin__ as builtin
-        >>> builtin.Some = Some
-
-    Next lets see if we can create an instance, pickle it and unpickle::
-
-        >>> some = Some('abc', 10)
-        >>> some.foo, some.bar
-        ('abc', 10)
-
-        >>> from pickle import dumps, loads
-        >>> some2 = loads(dumps(some))
-
-        >>> some2.foo, some2.bar
-        ('abc', 10)
-
-    """
-
-    __slots__ = []
-
-    def __getstate__(self, cls=None):
-        if cls is None:
-            # This is the case for the instance that gets pickled
-            cls = self.__class__
-
-        d = {}
-
-        # Get all data that should be stored from super classes
-        for c in cls.__bases__:
-            if hasattr(c, "__getstate__"):
-                d.update(c.__getstate__(self, c))
-
-        # Get all information that should be stored from cls and return the dict
-        for name in cls.__slots__:
-            if hasattr(self, name):
-                d[name] = getattr(self, name)
-
-        return d
-
-    def __setstate__(self, d):
-        # All values that were pickled are now assigned to a fresh instance
-        for name, value in d.iteritems():
-            try:
-                setattr(self, name, value)
-            except AttributeError:    # This is needed in cases like Rational :> Half
-                pass
-
 class Basic(object):
     """
     Base class for all objects in sympy.
